[DUOS-1950] renamed file to be CapitalCamelCase, added enum for FormFieldTypes
